Split data retrieval into several requests
I've increased the speed of gitcalendar twofold, according to Firebug, by
separating repo and commit retrieval into separate requests to the server. This
allows progressive loading of the calendar's marks without a more complex
streaming solution. Although it's still a bit slow to finish, the marks start
appearing much more quickly now. Not the prettiest solution, as it increases
the number of HTTP requests, but it greatly increases gitcalendar's usability.