Reworked caching to reflect view changes
In rewriting the data views, I had removed caching of the entire commit_list.
Now caching is back in place, for the repo list and each repo's commit list.
To avoid problems with a repo named "repos", I used a key of "_repos" for the
repo list cache. If this is a problem for you, please let me know.